Added new test refactor test to new location wrote a new test
