+++ .pre-commit-config.yaml
@@ -14,3 +14,4 @@ repos:
     rev: v2.15.3
     hooks:
     -   id: pylint
+        exclude: /
